`filterIt()` no longer calls `filterArray()`
