+++ src/transformation/index.ts
@@ -1,3 +1,4 @@
+export * from './filter-array';
 export * from './filter-it';
 export * from './flat-map-it';
 export * from './map-it';
